Serialize multiple Resources to a single Avro container file in HDFS
Implement two versions of SerializationFormat which serializes multiple Spring Resources
to a single Avro container file in HDFS. This is useful if we want to store a large number
of files in HDFS, and do so without hitting the NameNode memory limits.

Two different approaches are implemented since I'm not able to objectively assess their
pros-cons. And need more feedback/comments... :)